- Renamed the database folder.
